requirements.txt
numba~=0.58.1
numpy~=1.26.1
pandas~=2.1.1
matplotlib~=3.8.0
opencv-python~=4.8.1.78
scikit-learn~=1.3.1
stable-baselines3[extra]
ta-lib
binance_data
python-binance
gym~=0.26.2
gymnasium~=0.29.1
mplfinance~=0.12.10b0
tindicators~=0.9.12
finta~=1.3
ta~=0.10.2
websocket-client~=1.6.4
pympler~=1.0.1
pymoo~=0.6.0.1
python-dateutil~=2.8.2
scipy~=1.11.3
requests~=2.31.0
torch~=2.1.0+cu12
ccxt~=4.0.86
tensorflow~=2.14.0
keras-rl